refactor: move NamedTuple to data_types
